feat!: Consolidate array methods into `ts/array` block
- Also update readme for previous change
