+++ src/ts/ipv4-address.ts
@@ -1,5 +1,5 @@
-import { Err, Ok, type Result } from './result';
 import { isNumber } from './is-number';
+import { Err, Ok, type Result } from './result';
 
 type Octets = [number, number, number, number];
