--- src/ts/map-to-array.ts
@@ -1,26 +1,0 @@
-/** Maps the provided map into an array using the provided mapping function.
- *
- * @param map Map to be entered into an array
- * @param fn A mapping function to transform each pair into an item
- * @returns
- *
- * ## Example
- * ```ts
- * console.log(map); // Map(5) { 0 => 5, 1 => 4, 2 => 3, 3 => 2, 4 => 1 }
- *
- * const arr = mapToArray(map, (_, value) => value);
- *
- * console.log(arr); // [5, 4, 3, 2, 1]
- * ```
- */
-const mapToArray = <K, V, T>(map: Map<K, V>, fn: (key: K, value: V) => T): T[] => {
-	const items: T[] = [];
-
-	for (const [key, value] of map) {
-		items.push(fn(key, value));
-	}
-
-	return items;
-};
-
-export { mapToArray };
